abi method
