+++ src/core/abi/arc56-contract.ts
@@ -1,237 +1,3 @@
-import { Expand } from '../expand'
-import { ABIArgumentType, ABIMethod, ABIReturnType } from './abi-method'
-import { decodeABIValue, encodeABIValue, stringToABIType } from './abi-type'
-import { ABIValue } from './abi-value'
-import { ABITupleType, decodeTuple, encodeTuple } from './types'
-
-/** Type to describe an argument within an `Arc56Method`. */
-export type Arc56MethodArg = Expand<
-  Omit<Method['args'][number], 'type'> & {
-    type: ABIArgumentType
-  }
->
-
-/** Type to describe a return type within an `Arc56Method`. */
-export type Arc56MethodReturnType = Expand<
-  Omit<Method['returns'], 'type'> & {
-    type: ABIReturnType
-  }
->
-
-/**
- * Wrapper around `ABIMethod` that represents an ARC-56 ABI method.
- */
-export type Arc56Method = ABIMethod & {
-  args: Array<Arc56MethodArg>
-  returns: Arc56MethodReturnType
-}
-
-/**
- * Returns the `ABITupleType` for the given ARC-56 struct definition
- * @param struct The ARC-56 struct definition
- * @returns The `ABITupleType`
- */
-export function getABITupleTypeFromABIStructDefinition(struct: StructField[], structs: Record<string, StructField[]>): ABITupleType {
-  const childTypes = struct.map((v) =>
-    typeof v.type === 'string'
-      ? structs[v.type]
-        ? getABITupleTypeFromABIStructDefinition(structs[v.type], structs)
-        : stringToABIType(v.type)
-      : getABITupleTypeFromABIStructDefinition(v.type, structs),
-  )
-  return {
-    name: 'Tuple',
-    childTypes: childTypes,
-  } satisfies ABITupleType
-}
-
-/**
- * Converts a decoded ABI tuple as a struct.
- * @param decodedABITuple The decoded ABI tuple value
- * @param structFields The struct fields from an ARC-56 app spec
- * @returns The struct as a Record<string, any>
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function getABIStructFromABITuple<TReturn extends ABIStruct = Record<string, any>>(
-  decodedABITuple: ABIValue[],
-  structFields: StructField[],
-  structs: Record<string, StructField[]>,
-): TReturn {
-  return Object.fromEntries(
-    structFields.map(({ name: key, type }, i) => {
-      // TODO: confirm if we need this logic
-      // let abiType: ABIType
-
-      // if (typeof type === 'string') {
-      //   if (type in structs) {
-      //     abiType = getABITupleTypeFromABIStructDefinition(structs[type], structs)
-      //   } else {
-      //     abiType = stringToABIType(type)
-      //   }
-      // } else {
-      //   abiType = getABITupleTypeFromABIStructDefinition(type, structs)
-      // }
-
-      // const abiValue = convertAbiByteArrays(decodedABITuple[i], abiType)
-      // const convertedValue = convertABIDecodedBigIntToNumber(abiValue, abiType)
-      return [
-        key,
-        (typeof type === 'string' && !structs[type]) || !Array.isArray(decodedABITuple[i])
-          ? decodedABITuple[i]
-          : getABIStructFromABITuple(decodedABITuple[i], typeof type === 'string' ? structs[type] : type, structs),
-      ]
-    }),
-  ) as TReturn
-}
-
-/**
- * Converts an ARC-56 struct as an ABI tuple.
- * @param struct The struct to convert
- * @param structFields The struct fields from an ARC-56 app spec
- * @returns The struct as a decoded ABI tuple
- */
-export function getABITupleFromABIStruct(
-  struct: ABIStruct,
-  structFields: StructField[],
-  structs: Record<string, StructField[]>,
-): ABIValue[] {
-  return structFields.map(({ name: key, type }) => {
-    const value = struct[key]
-    return typeof type === 'string' && !structs[type]
-      ? (value as ABIValue)
-      : getABITupleFromABIStruct(value as ABIStruct, typeof type === 'string' ? structs[type] : type, structs)
-  })
-}
-
-/** Decoded ARC-56 struct as a struct rather than a tuple. */
-export type ABIStruct = {
-  [key: string]: ABIStruct | ABIValue
-}
-
-/**
- * Returns the decoded ABI value (or struct for a struct type)
- * for the given raw Algorand value given an ARC-56 type and defined ARC-56 structs.
- * @param value The raw Algorand value (bytes or uint64)
- * @param type The ARC-56 type - either an ABI Type string or a struct name
- * @param structs The defined ARC-56 structs
- * @returns The decoded ABI value or struct
- */
-export function getABIDecodedValue(
-  value: Uint8Array | number | bigint,
-  type: string,
-  structs: Record<string, StructField[]>,
-): ABIValue | ABIStruct {
-  if (type === 'AVMBytes' || typeof value !== 'object') return value
-  if (type === 'AVMString') return Buffer.from(value).toString('utf-8')
-  if (type === 'AVMUint64') return decodeABIValue(stringToABIType('uint64'), value)
-  if (structs[type]) {
-    const tupleValue = decodeTuple(getABITupleTypeFromABIStructDefinition(structs[type], structs), value)
-    return getABIStructFromABITuple(tupleValue, structs[type], structs)
-  }
-
-  // TODO: do we need this logic?
-  // const abiType = stringToABIType(type)
-  // const decodedValue = convertAbiByteArrays(abiType.decode(value), abiType)
-  // return convertABIDecodedBigIntToNumber(decodedValue, abiType)
-
-  const abiType = stringToABIType(type)
-  return decodeABIValue(abiType, value)
-}
-
-/**
- * Returns the ABI-encoded value for the given value.
- * @param value The value to encode either already in encoded binary form (`Uint8Array`), a decoded ABI value or an ARC-56 struct
- * @param type The ARC-56 type - either an ABI Type string or a struct name
- * @param structs The defined ARC-56 structs
- * @returns The binary ABI-encoded value
- */
-export function getABIEncodedValue(
-  value: Uint8Array | ABIValue | ABIStruct,
-  type: string,
-  structs: Record<string, StructField[]>,
-): Uint8Array {
-  if (typeof value === 'object' && value instanceof Uint8Array) return value
-  if (type === 'AVMUint64') return encodeABIValue(stringToABIType('uint64'), value as bigint | number)
-  if (type === 'AVMBytes' || type === 'AVMString') {
-    if (typeof value === 'string') return Buffer.from(value, 'utf-8')
-    if (typeof value !== 'object' || !(value instanceof Uint8Array)) throw new Error(`Expected bytes value for ${type}, but got ${value}`)
-    return value
-  }
-  if (structs[type]) {
-    const tupleType = getABITupleTypeFromABIStructDefinition(structs[type], structs)
-    if (Array.isArray(value)) {
-      encodeTuple(tupleType, value)
-    } else {
-      const values = getABITupleFromABIStruct(value as ABIStruct, structs[type], structs)
-      return encodeTuple(tupleType, values)
-    }
-  }
-
-  const abiType = stringToABIType(type)
-  return encodeABIValue(abiType, value as ABIValue)
-}
-
-/**
- * Returns the ARC-56 ABI method object for a given method name or signature and ARC-56 app spec.
- * @param methodNameOrSignature The method name or method signature to call if an ABI call is being emitted.
- * e.g. `my_method` or `my_method(unit64,string)bytes`
- * @param appSpec The app spec for the app
- * @returns The `Arc56Method`
- */
-export function getArc56Method(methodNameOrSignature: string, appSpec: Arc56Contract): Arc56Method {
-  let method: Method
-  if (!methodNameOrSignature.includes('(')) {
-    const methods = appSpec.methods.filter((m) => m.name === methodNameOrSignature)
-    if (methods.length === 0) throw new Error(`Unable to find method ${methodNameOrSignature} in ${appSpec.name} app.`)
-    if (methods.length > 1) {
-      throw new Error(
-        `Received a call to method ${methodNameOrSignature} in contract ${
-          appSpec.name
-        }, but this resolved to multiple methods; please pass in an ABI signature instead: ${appSpec.methods
-          .map((m) => new ABIMethod(m).getSignature())
-          .join(', ')}`,
-      )
-    }
-    method = methods[0]
-  } else {
-    const m = appSpec.methods.find((m) => new ABIMethod(m).getSignature() === methodNameOrSignature)
-    if (!m) throw new Error(`Unable to find method ${methodNameOrSignature} in ${appSpec.name} app.`)
-    method = m
-  }
-  return new Arc56Method(method)
-}
-
-/**
- * Checks for decode errors on the AppCallTransactionResult and maps the return value to the specified generic type
- *
- * @param returnValue The smart contract response
- * @param method The method that was called
- * @param structs The struct fields from the app spec
- * @returns The smart contract response with an updated return value
- */
-export function getArc56ReturnValue<TReturn extends Uint8Array | ABIValue | ABIStruct | undefined>(
-  returnValue: ABIReturn | undefined,
-  method: Method | Arc56Method,
-  structs: Record<string, StructField[]>,
-): TReturn {
-  const m = 'method' in method ? method.method : method
-  const type = m.returns.struct ?? m.returns.type
-  if (returnValue?.decodeError) {
-    throw returnValue.decodeError
-  }
-  if (type === undefined || type === 'void' || returnValue?.returnValue === undefined) return undefined as TReturn
-
-  if (type === 'AVMBytes') return returnValue.rawReturnValue as TReturn
-  if (type === 'AVMString') return Buffer.from(returnValue.rawReturnValue).toString('utf-8') as TReturn
-  if (type === 'AVMUint64') return ABIType.from('uint64').decode(returnValue.rawReturnValue) as TReturn
-
-  if (structs[type]) {
-    return getABIStructFromABITuple(returnValue.returnValue as ABIValue[], structs[type], structs) as TReturn
-  }
-
-  return convertAbiByteArrays(returnValue.returnValue, ABIType.from(type)) as TReturn
-}
-
 /****************/
 /** ARC-56 spec */
 /****************/
@@ -260,7 +26,7 @@ export interface Arc56Contract {
   /** Named structs used by the application. Each struct field appears in the same order as ABI encoding. */
   structs: { [structName: StructName]: StructField[] }
   /** All of the methods that the contract implements */
-  methods: Method[]
+  methods: Arc56Method[]
   state: {
     /** Defines the values that should be used for GlobalNumUint, GlobalNumByteSlice, LocalNumUint, and LocalNumByteSlice when creating the application  */
     schema: {
@@ -332,7 +98,7 @@ export interface Arc56Contract {
   templateVariables?: {
     [name: string]: {
       /** The type of the template variable */
-      type: ABITypeFoo | AVMType | StructName
+      type: ABITypeName | AVMType | StructName
       /** If given, the base64 encoded value used for the given app/program */
       value?: string
     }
@@ -341,13 +107,13 @@ export interface Arc56Contract {
   scratchVariables?: {
     [name: string]: {
       slot: number
-      type: ABITypeFoo | AVMType | StructName
+      type: ABITypeName | AVMType | StructName
     }
   }
 }
 
 /** Describes a method in the contract. This interface is an extension of the interface described in ARC-4 */
-export interface Method {
+export interface Arc56Method {
   /** The name of the method */
   name: string
   /** Optional, user-friendly description for the method */
@@ -355,7 +121,7 @@ export interface Method {
   /** The arguments of the method, in order */
   args: Array<{
     /** The type of the argument. The `struct` field should also be checked to determine if this arg is a struct. */
-    type: ABITypeFoo
+    type: ABITypeName
     /** If the type is a struct, the name of the struct */
     struct?: StructName
     /** Optional, user-friendly name for the argument */
@@ -367,7 +133,7 @@ export interface Method {
       /** Base64 encoded bytes, base64 ARC4 encoded uint64, or UTF-8 method selector */
       data: string
       /** How the data is encoded. This is the encoding for the data provided here, not the arg type */
-      type?: ABITypeFoo | AVMType
+      type?: ABITypeName | AVMType
       /** Where the default value is coming from
        * - box: The data key signifies the box key to read the value from
        * - global: The data key signifies the global state key to read the value from
@@ -381,7 +147,7 @@ export interface Method {
   /** Information about the method's return value */
   returns: {
     /** The type of the return value, or "void" to indicate no return value. The `struct` field should also be checked to determine if this return value is a struct. */
-    type: ABITypeFoo
+    type: ABITypeName
     /** If the type is a struct, the name of the struct */
     struct?: StructName
     /** Optional, user-friendly description for the return value */
@@ -431,7 +197,7 @@ export interface Event {
   /** The arguments of the event, in order */
   args: Array<{
     /** The type of the argument. The `struct` field should also be checked to determine if this arg is a struct. */
-    type: ABITypeFoo
+    type: ABITypeName
     /** Optional, user-friendly name for the argument */
     name?: string
     /** Optional, user-friendly description for the argument */
@@ -442,7 +208,7 @@ export interface Event {
 }
 
 /** An ABI-encoded type */
-export type ABITypeFoo = string // TODO: find a better name
+export type ABITypeName = string
 
 /** The name of a defined struct */
 export type StructName = string
@@ -464,7 +230,7 @@ export interface StructField {
   /** The name of the struct field */
   name: string
   /** The type of the struct field's value */
-  type: ABITypeFoo | StructName | StructField[]
+  type: ABITypeName | StructName | StructField[]
 }
 
 /** Describes a single key in app storage */
@@ -472,10 +238,10 @@ export interface StorageKey {
   /** Description of what this storage key holds */
   desc?: string
   /** The type of the key */
-  keyType: ABITypeFoo | AVMType | StructName
+  keyType: ABITypeName | AVMType | StructName
 
   /** The type of the value */
-  valueType: ABITypeFoo | AVMType | StructName
+  valueType: ABITypeName | AVMType | StructName
   /** The bytes of the key encoded as base64 */
   key: string
 }
@@ -485,9 +251,9 @@ export interface StorageMap {
   /** Description of what the key-value pairs in this mapping hold */
   desc?: string
   /** The type of the keys in the map */
-  keyType: ABITypeFoo | AVMType | StructName
+  keyType: ABITypeName | AVMType | StructName
   /** The type of the values in the map */
-  valueType: ABITypeFoo | AVMType | StructName
+  valueType: ABITypeName | AVMType | StructName
   /** The base64-encoded prefix of the map keys*/
   prefix?: string
 }
